Relocate misplaced command helpers to their data's modules
Move three helpers out of the command layer to sit beside the data they
operate on, each with its tests:

- validate_unique_guest_paths -> config.rs (next to Mount)
- validate_git_repo_workspace_mounts -> workspace.rs (uses default_workspace_path)
- git_repo_default_instance_name -> github_repo.rs (derives from a repo slug)

validate_copy_workspace_mounts stays in lifecycle: it is coupled to the
command-layer ProjectTransport enum, so moving it would invert the
config/workspace -> commands dependency direction. No behavior change.
